refactor: util package 추가
